+++ BagOLoot/Actions/RemoveToy.cs
@@ -16,6 +16,13 @@ namespace BagOLoot.Actions
       {
           Console.WriteLine($"{Array.IndexOf(children,child)+1}. {child.name}");
       }
+      // var children =
+      //   from c in book
+      //   select c;
+      // foreach (var c in children)
+      // {
+      //   Console.WriteLine($"{c}");
+      // }
 
       Console.Write ("> ");
       string childChoice = Console.ReadLine();
